change code convention

+++ redisson-demo/src/main/java/com/haopn/demo/entity/Count.java
@@ -7,22 +7,22 @@ import javax.persistence.*;
 public class Count {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private int number;
+    private long id;
+    private long number;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getNumber() {
+    public long getNumber() {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(long number) {
         this.number = number;
     }
 }
