+++ redisson-demo/src/main/java/com/haopn/demo/repository/CountRepository.java
@@ -3,6 +3,6 @@ package com.haopn.demo.repository;
 import com.haopn.demo.entity.Count;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CountRepository extends JpaRepository<Count, Integer> {
-    Count findById(int id);
+public interface CountRepository extends JpaRepository<Count, Long> {
+    Count findById(long id);
 }
